client: give the server a sec before signaling it with SIGINT
When the client exits successfully, it signals the server with SIGINT.
Before changing the way the event loop is managed, the server always
exited before being signaled during integrations tests.

Now though, it seems that the server exits slightly after, so it does
often receive the SIGINT during integration tests. When it does, it
prints a KeyboardInterrupt backtrace to its standard error stream.
Integration tests treat this as a failure.

Adding a sleep of 0.1 second before signaling the server with SIGINT
seems to fix the issue. Just to be sure, we are waiting up to one second
before sending the SIGINT.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>