Keep app.variant consistent with app.base_model.source.current.variant
Otherwise you end up in this situation where GET /meta/client_variant
returns a different value to the variant you'd get from GET /source, and
that's just confusing. One way this confusion shows up is that if the
default source in the catalog is not server it is not selected, because
some slightly confusing client-server interactions end up setting the
entry on the source model to one of the synthesized ones.

Perhaps these don't need to be two separate attributes, but I think
system_setup depends on app.variant so leave that bit for now.