ssh: fetch SSH keys on the server side and expose API for it
Instead of fetching SSH keys on the client side, we now make the client
consume an API and have the implementation on the server.

The main benefit is that it gives us more control over the environment
where the ssh-import-id command is executed.

This should allow us to set HTTP proxy environment variables (and
optionally locale-related variables such as LC_MESSAGES) according to
the user's selection.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>