ssh: use specified proxy when importing SSH keys
When fetching SSH keys, the proxy settings specified by the user were
not used. This resulted in the inability to import keys in networks where
a HTTP proxy is mandatory.

We now explicitly set the https_proxy environment variable when calling
ssh-import-id if a proxy was configured by the user.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>