ubuntu-pro: cancel any contract selection before initiating a new one
Upon reaching the Ubuntu Pro screen that shows the user-code, we need to
initiate a contract selection. If we are navigating to the screen a
second time though, we need to cancel existing contract selections
before initiating a new one.

Failing to do so results in a UPCSAlreadyInitiatedError.

(cherry picked from commit 0736885d46fdb29a36b5c1cc528e025285b23010)
Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>