drivers: don't mark configured automatically, the client decides
The drivers controller used to mark the model configured automatically
if no drivers were found. This is okay in most cases but sometimes, we
want to query again the list of drivers (after the client_variant gets
set for instance) and we don't want the model to remain configured.

Let's leave this up to the client to decide if the model should be
configured.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>