refresh: remove the cancel update button since it is broken
Cancelling the installer self refresh does not work. Clicking on the
button simply goes back to the screen asking the user if they want to
update ; but the refresh operation still continues in the background.

Since we do not have a simple way to implement refresh cancellation at
the moment, we are just removing the button.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>