kvm-test: add option to add extra disks for install
The patch introduces a new option --disk-count to create additional disks in
kvm-test. It also makes the --disksize option usable multiple times so we can
specify the size of each disk.

 $ kvm-test.py --disk-count 2

... will result in two disks, each of the default size (12G)

 $ kvm-test.py --disk-count 2 --disk-size 5G --disk-size 10G

... will result in a disk of size 5G and another disk of size 10G.

For now, this approach does not allow to selectively specify the size of each
disk (i.e., you can't specify the size of disk 2 without specifying the size of
disk 1).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>