mirror: add new endpoint to submit mirror without marking configured
The POST handler for /mirror applies the URL passed but also marks the
model configured. That said, we want the ability to:

 1. supply a mirror URL
 2. run the mirror test
 3. mark the model configured if the test is successful

This patches creates a new endpoint: /mirror/candidate that does the
same as /mirror except that it does not mark the model configured.

It is therefore suitable for step 1.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>