storage: fix creation of encrypted VG in manual partitioning
The following patch changed the name of the luks passphrase field from
"password" to "passphrase" to make it consistent across the screens:

  commit d63b44c014ebaa2831f683f7397cdf141e2641b4
  storage: use fields named passphrase for passphrases

  Because the storage views lean on the implementation of
  setup_password_validation from the identity screen, we were forced to
  use a form with fields named "password" and "password_confirm".

  This makes the code confusing because we use the "passphrase"
  terminology in the storage forms.

  We now leave up to the caller to specify which fields he wants to be
  part of the validation ; instead of making him provide the full form.

However, this change broke the ability to create an encrypted VG in
manual partitioning mode.

This happened because the values input in the VG edit form are passed
as-is (mostly) to functions that are also used by the server.

On one hand, we have the client which deals with passphrases named
"passphrase" and "confirm_passphrase".
On the other hand, we have the server which deals with passphrases
named "password" and "confirm_password".

This inconsistency makes it hard to work with shared code.

To work around the issue, we now rename the passphrase key (passphrase
-> password) before passing it to the shared code.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>