mirror: only donwload index file for mirror testing
On slow connections, running a full apt-get update for the sole purpose
of mirror testing ; and discarding the result is a bad thing to do.

This can take minutes and consume over 50 MiB.

Instead, we use mwhudson's approach and only download the index files.
Instructing apt to do so is a bit clunky but it seems to be worth the
effort.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>