storage: use "passphrase" for LVM/LUKS in the server
The client expected the LUKS passphrase to be called "passphrase" but
the server expected it to be called "password".

To keep the implementation consistent, we now use "passphrase"
everywhere except in the API (i.e., storage/guided and
storage/v2/guided) where "password" is still used for backward
compatibility reasons.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>