mounter: remove conflicting instructions about sources.list.d
In dry-run mode, when generating the fake overlay for apt, we copy the
content of /etc/apt from the host to the temporary directory.
For some reason, we then removed the sources.list.d directory in the
destination.

When I noticed the absence of sources.list.d in the fake overlay, I
added another call to cp(1) to get it populated. However, a more
sensible thing to do is to get rid of the instruction that removed
sources.list.d.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>