apt: use wording final instead of elected to reduce coupling with mirror
The mirror model has the notion of primary candidates, staged primary
candidate and elected primary candidate. This is purely specific to the
mirror model and does not necessarily make sense from the outside.

In code that does not specifically need to know implmentation details of
the mirror model, we now prefer the wording "final configuration" over
"elected configuration".

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>