ssh: don't run ssh-import-id as part of the CI
Running ssh-import-id as part of the CI has limitations. First it
requires that we have access to the Internet. Second, it is affected by
github and launchpad service disruptions and or rate limiting policies.

Make sure we don't call ssh-import-id as part of the CI. Instead use a
username that is configured to produce fake successful key imports.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>