source: honor source if hinted by autoinstall but interactive
If a specific source ID is set via autoinstall but the source controller
is marked interactive, the UI would select the default source (i.e., the
first one on the list) instead of the one matching the ID specified.

This happens because the matching between the ID and the sources loaded
was done in apply_autoinstall_config ; which is only called in
non-interactive scenarios.

By moving this matching in the start coroutine instead, the right source
is also selected when the source controller is interactive.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>