mirror: pass an alternative config file to apt through the environment
Sadly, passing an alternative config file to apt using -o CLI options
does not work because apt:

1) Reads the configuration file first
2) Reads CLI options (and override specified settings) second

Therefore, the default configuration file is always read, instead of the
one we supply.

To specify an alternative configuration file, the recommendation from
the apt maintainer is to supply an APT_CONFIG variable, hinting apt to
read said file instead of the default.

We now generate a temporary file, write the directives we need to it
using the python apt library, and then execute apt-get with a path to
this temporary file set in APT_CONFIG.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>