apt: force downloads to be unsandboxed when running mirror testing
When apt is run as root, APT tries to drop privileges by setuid-ing to
the APT::Sandbox::User user (i.e., _apt by default). This does not work
for us when doing mirror testing because the default sandbox user does
not have access to the overlay. Therefore, APT produces a warning and
reverts to unsandboxed downloads.

Let's force apt to use unsandboxed downloads by setting root as the
sandbox user. This has the same result but avoids showing a warning in
the APT output during mirror testing.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>