storage: lengthen and log probing time
A reasonably common type of bug report is hitting the block probing
timeout.  That timeout is at 15 seconds today.  Extend that to 90
seconds, and log the time it takes.

As a performance exercise, it would be good to know where that time is
being spent and why it takes 10x longer or more for some people.