filesystem: ensure probe wait handle start timing
In LP: 2009797, an exception of this form happens:
AttributeError: 'FilesystemController' object has no attribute '_start_task'

The installer client, u-d-i, is asking for storage information ASAP
after the socket starts listening, and in this case that happened before
all controllers were started.  The sync primitive the probe is waiting
on wasn't created yet.

Add an event earlier in sequence to ensure that the probe has something
to wait on, even before the controller is officially `start()`ed.