filesystem: fix logical+swap or logical+boot partition not considered logical
Some people who installed a previous version of Ubuntu on a DOS
partition table ended up with a logical, ESP partition (which is marked
bootable).

This is an unsupported use-case. An ESP partition should always be a
primary partition. Ideally, an ESP partition should be on a GPT.

However, curtin's storage_config makes it clear that expecting the flag
of a logical partition to be set to 'logical' is a bad idea.

  # if the boot flag is set, use this as the flag, logical
  # flag is not required as we can determine logical via
  # partition number

Other people want to create a swap partition in an extended partition.
This is a perfectly valid use-case and we set flag='swap' for those
partitions.

Therefore, in subiquity code, we cannot just check if a partition has a
flag set to 'logical'. We need to look at the partition number too.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>