simplify InstallController.curtin_install a bit
curtin_install before this assembled a list of steps and then executed
them one by one. As far as I can tell/remember (I was at least partly
responsible for this) there is no reason to do it this way: it feels
more Pythonic to just execute the steps!

This required removing some indirection in how the config for each step
was constructed, which I think also makes things a bit clearer.