ubuntu-drivers: by default in dry-run, consider that drivers are available
We only search for drivers if search_drivers is enabled in the source
screen. When testing in dry-run mode, we need to check the
search_drivers checkbox AND remember to pass the has-drivers debug flag.
I think it makes sense to consider drivers available by default in
dry-run mode.

This also means that OEM meta-packages are available by default (and
_should_ be installed automatically) in dry-run now.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>