drivers: fix empty list of drivers shown when it has not yet been retrieved
The /drivers GET endpoint is supposed to have {"drivers": null} when the
list of drivers has not (yes) been retrieved.

However, this behavior got broken by the following patch:

 66c82b322 drivers: query again list of drivers if the source variant changes

.. which resets the drivers variable to an empty list when querying the
list of drivers (this happens automatically when the user submits on the
source screen).

Therefore, instead of getting {"drivers": null}, we would get
{"drivers": []} which has a different meaning.

The patch also incorrectly declared the drivers variable in the
initializer (not taking None into account). The variable was actually
already declared as a class variable with the right type.

Fixed by resetting the variable to None when re-querying the list of
drivers and dropped the class variable in favor of the instance variable
(with the right type this time).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>