apt: reset only diverted files on deconfigure
Before starting the actual installation, we make changes to etc/apt
directory in an overlay so that the packages are pulled from the media
instead of the network.

At the end of installation, we need to revert these changes. We used to
replace the etc/apt directory on the target by the etc/apt directory
from the configured overlay.

This is a simple way to revert the temporary changes. Unfortunately, it
also removes configuration files that extra packages may have installed
in etc/apt. For instance, when installing an OEM meta-package, a
etc/apt/sources.list.d/<oem-meta-package>.list gets installed.

If we reset etc/apt to its content from the configured overlay, we lose
this file.

Fixed by selectively discarding changes to the files that subiquity
modified.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>