kernel: add a way to override the kernel metapkg we want to install
Before running curtin's curthooks, we use the kernel model to configure
which kernel meta-package we want to install on the target system.

There are multiple ways to select the kernel flavor, including
autoinstall directives or placing a file in
/etc/subiquity/kernel-meta-package.

That said, if we install Ubuntu on certified hardware, the OEM
meta-package will have something to say about the expected kernel
flavor. When installing the kernel meta-package, it will automatically
pull the kernel that it expects to run on.

Therefore, we should make it possible to override the kernel flavor so
that we don't end up with multiple kernels installed.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>