oem: check if OEM meta-packages want OEM or default kernel flavor
When installing on certified hardware, we now look for the
Ubuntu-Oem-Kernel-Flavour attribute of the OEM meta-package to determine
which kernel flavour we should use. This attribute can take two values:

 * oem -> in which case we make use of the override to install the
          OEM kernel flavor
 * default -> in which case we get the kernel flavor that was originally
   configured.

If the attribute is not present or contains any other value, it is
treated the same as if it was "oem".

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>