oem: skip listing OEM metapkgs on ubuntu-server
OEM meta-packages have an attribute stating if they want the OEM kernel
or the "default" kernel to be installed alongside them.

That said, in the OEM archive, they expect the "default" OEM kernel to
be the HWE kernel. This is only true for ubuntu-desktop. Let's not
install OEM meta-packages on ubuntu-server for now.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>