have v2_guided_GET indicate when manual partitioning is possible
There are some install scenarios where manual partitioning is not
possible at all. The current way we install core boot classic is one
(although that might change as and when we get partial gadgets) but
other situations might be when doing an install that just creates a
reset partition, or when installing a dd image rather than a squashfs.