rewrite guided disk selection to be more general
Drive various decisions that used to be keyed off core boot capabilities
and special purpose bits of the API by the new guided capability stuff.
I tried to think of a way to do this incrementally and failed. It might
be easier to review the new code rather than the diff in places.